requirements.txt
django==4.1
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
pep8-naming==0.13.2
django-debug-toolbar==3.2.4
django-crispy-forms==1.14.0
asgiref==3.5.2
djangorestframework==3.13.1
pytz==2022.2.1
sqlparse==0.4.2
